Add files via upload
This project is developed using Html, css and java script, Bootstrap for the frontend and Python Django for the backend. Naïve Bayes algorithm is implemented with NLTK which is used in  the machine learning part. Accuracy of this model is very high and website is very responsive for the user inputs. Django is very stable for backend web development this will provide best user experience and capable of handling the load.